requirements.txt
ez_setup==0.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==74.1.1
